requirements.txt
azure-functions
azure-identity
azure-mgmt-web
azure-monitor-query
requests
ipaddress
